Update game_of_life.py
It seems that the last parts were not committed by sourcetree. Here I updated the description, I found out the bounds for evolve were wrong and needed to be from range of the length. I also made it so that I can get an image from the last iteration and I also got the code for the command line from chatgpt.